# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MentorAlm API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new student account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/signin:
    post:
      operationId: signin
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SigninInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token using refresh token cookie
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Send OTP to email for password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking and Razorpay order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingInput"
      responses:
        "201":
          description: Booking initiated with Razorpay order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingOrderResponse"
  /bookings/verify:
    post:
      operationId: verifyBooking
      tags: [bookings]
      summary: Verify Razorpay payment and confirm booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentInput"
      responses:
        "200":
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/mine:
    get:
      operationId: getMyBookings
      tags: [bookings]
      summary: Get current user bookings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingWithService"
...
  /admin/bookings:
    get:
      operationId: listAdminBookings
      tags: [admin]
      summary: List all bookings
      security:
        - bearerAuth: []
      parameters:
        - name: serviceId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminBooking"
  /admin/bookings/{id}:
    patch:
      operationId: updateAdminBooking
      tags: [admin]
      summary: Update booking status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusInput"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBooking"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    SignupInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 6
        confirmPassword:
          type: string
      required:
        - name
        - email
        - password
        - confirmPassword
    SigninInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    ForgotPasswordInput:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    ResetPasswordInput:
      type: object
      properties:
        email:
          type: string
          format: email
        otp:
          type: string
        newPassword:
          type: string
          minLength: 6
      required:
        - email
        - otp
        - newPassword
    UserPublic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        user:
          $ref: "#/components/schemas/UserPublic"
      required:
        - accessToken
        - user
    RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
      required:
        - accessToken
...
    Service:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        shortDesc:
          type: string
        fullDesc:
          type: string
        included:
          type: string
        category:
          type: string
        duration:
          type: string
        price:
          type: integer
        counsellorName:
          type: string
        counsellorBio:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        status:
          type: string
        slots:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - shortDesc
        - fullDesc
        - included
        - category
        - duration
        - price
        - counsellorName
        - counsellorBio
        - status
        - slots
        - createdAt
...
    CreateBookingInput:
      type: object
      properties:
        serviceId:
          type: integer
        slotDateTime:
          type: string
      required:
        - serviceId
        - slotDateTime
    VerifyPaymentInput:
      type: object
      properties:
        bookingId:
          type: integer
        razorpayOrderId:
          type: string
        razorpayPaymentId:
          type: string
        razorpaySignature:
          type: string
      required:
        - bookingId
        - razorpayOrderId
        - razorpayPaymentId
        - razorpaySignature
    Booking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        serviceId:
          type: integer
        slotDateTime:
          type: string
        paymentStatus:
          type: string
        razorpayOrderId:
          type: string
          nullable: true
        razorpayPaymentId:
          type: string
          nullable: true
        amount:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - serviceId
        - slotDateTime
        - paymentStatus
        - amount
        - status
        - createdAt
    BookingWithService:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        serviceId:
          type: integer
        slotDateTime:
          type: string
        paymentStatus:
          type: string
        razorpayOrderId:
          type: string
          nullable: true
        razorpayPaymentId:
          type: string
          nullable: true
        amount:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        service:
          $ref: "#/components/schemas/Service"
      required:
        - id
        - userId
        - serviceId
        - slotDateTime
        - paymentStatus
        - amount
        - status
        - createdAt
        - service
    BookingOrderResponse:
      type: object
      properties:
        bookingId:
          type: integer
        razorpayOrderId:
          type: string
        razorpayKeyId:
          type: string
        amount:
          type: integer
        currency:
          type: string
      required:
        - bookingId
        - razorpayOrderId
        - razorpayKeyId
        - amount
        - currency
...
    AdminBooking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        studentName:
          type: string
        studentEmail:
          type: string
        serviceId:
          type: integer
        serviceTitle:
          type: string
        slotDateTime:
          type: string
        paymentStatus:
          type: string
        amount:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - studentName
        - studentEmail
        - serviceId
        - serviceTitle
        - slotDateTime
        - paymentStatus
        - amount
        - status
        - createdAt
    UpdateBookingStatusInput:
      type: object
      properties:
        status:
          type: string
      required:
        - status
...