# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MentorAlm API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  userId: number;
  serviceId: number;
  slotDateTime: string;
  paymentStatus: string;
  razorpayOrderId?: string | null;
  razorpayPaymentId?: string | null;
  amount: number;
  status: string;
  createdAt: Date;
}